optimize plot memory managment to keep more plots in memory
Former-commit-id: 01c4a0d8685702b3c6e72dc27c9db48661ab14fd
Former-commit-id: 3c926506f72bed2183d94dd5782452d4a2a59d64

+++ java/src/edu/caltech/ipac/firefly/server/visualize/VisServerOps.java
@@ -64,7 +64,6 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.RandomAccessFile;
 import java.util.ArrayList;
-import java.util.ConcurrentModificationException;
 import java.util.HashMap;
 import java.util.List;
 
@@ -175,27 +174,7 @@ public class VisServerOps {
 
     public static boolean deletePlot(String ctxStr) {
         PlotClientCtx ctx= VisContext.getPlotCtx(ctxStr);
-        if (ctx!=null) {
-            VisContext.deletePlotCtx(ctx);
-            File delFile;
-            List<PlotImages> allImages= ctx.getAllImagesEveryCreated();
-            try {
-                for(PlotImages images : allImages) {
-                    for(PlotImages.ImageURL image : images) {
-                        delFile=  VisContext.convertToFile(image.getURL());
-                        delFile.delete(); // if the file does not exist, I don't care
-                    }
-                    String thumbUrl= images.getThumbnail()!=null ? images.getThumbnail().getURL(): null;
-                    if (thumbUrl!=null) {
-                        delFile=  VisContext.convertToFile(images.getThumbnail().getURL());
-                        delFile.delete(); // if the file does not exist, I don't care
-                    }
-                }
-            } catch (ConcurrentModificationException e) {
-                // just abort, we can get it next time
-            }
-
-        }
+        if (ctx!=null)  VisContext.deletePlotCtx(ctx);
         return true;
     }
